feat: Implementar upgrade de arma
