+++ Assets/Prefabs/kidZombie.prefab
@@ -18,7 +18,7 @@ GameObject:
   - component: {fileID: 1473362628178941878}
   m_Layer: 11
   m_Name: kidZombie
-  m_TagString: Untagged
+  m_TagString: Enemy
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
