+++ Assets/Prefabs/skinnyZombie.prefab
@@ -18,7 +18,7 @@ GameObject:
   - component: {fileID: 6355506583925365000}
   m_Layer: 11
   m_Name: skinnyZombie
-  m_TagString: Untagged
+  m_TagString: Enemy
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
