+++ Assets/Prefabs/zombie.prefab
@@ -18,7 +18,7 @@ GameObject:
   - component: {fileID: 8193606659908256258}
   m_Layer: 11
   m_Name: zombie
-  m_TagString: Untagged
+  m_TagString: Enemy
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
