--- Assets/Scripts/Bullet.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 524fc856945bd0a45a82593776009490
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
