+++ Assets/Scripts/Purchase.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9c58f74d6c7d49143a6a00f6e4fa471a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
