+++ Assets/Scripts/Purchase/GatePurchase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5203712f8d83f614f94532659caa1c47
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
