+++ Assets/Scripts/Purchase/GunPurchase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8d033b0b0b2f8d34ca867566b2c34bb7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
