+++ Assets/Scripts/Purchase/UpgradePurchase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2017b71fdefacc0418efc90799f2bd17
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
